move library wrappers to lib module
